Add Qt::InputMethodHint::IhmNoFullscreen to affect IM sizing
On Android, the change 1f6d7cbb341bd79826d3f6d69e1f1a427ebb8f1b removed
the imeOptions.IME_FLAG_NO_EXTRACT_UI flag that suppressed the
fullscreen IME with extracted view covering the whole display on
landscape mode. Currently when to apply fullscreen IME is up to Android
and not forced nor suppressed by Qt. It looks like the default behavior
now in most cases at least is that the IME opens in fullscreen in
landscape mode.

Having the keyboard open in fullscreen is not always wanted and having a
way to disable it is really needed in some cases e.g a scenario where
the UI is designed to filter and show UI elements dynamically based on
what is written in the text field, in this case the user can't see
what's changing before closing the IME.

This change keeps the default behavior of allowing Android to decide
when to show fullscreen IME while but adding the possibility to disable
it completely when needed/wanted. Note that even with the
ImhNoFullscreen flag set, it's up to the IME in use to respect it or
not, meaning that the behavior is not guaranteed.

This change is targeted only towards Android. It seems iOS does not have
a fullscreen keyboard mode at all so this kind of change is not
applicable. OpenHarmony OS does have a fullscreen keyboard possibility,
it is controlled with inputmethodengine.DISPLAY_MODE_FULL and
inputmethodengine.EnhancedPanelRect.fullscreenMode, but these are not
possible to control from applications.

[ChangeLog][QtGui] Added Qt::InputMethodHint::IhmNoFullscreen to request
that the platform should not make input method panels fullscreen.

Task-number: QTBUG-143161
Change-Id: If411f43e71e48573872f9670967b7b1b0c3a67df
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>