JNI: move exception-handling logic into LocalFrame subclass
QJniObject is whole-sale exported, so the private constructor that takes
a LocalFrame generates a symbol. "nm" shows that this symbol is changing
between Qt 6.10 and 6.11, as a result of the LocalFrame template getting
an explicit template parameter for the return type. This might result in
binary compatibility issues for existing projects.

The use of LocalFrame in this particular context doesn't require an
explicit return type to be specified, so the "old" LocalFrame is enough
here. To keep the old symbol unchanged, add the new exception handling
facilities to a LocalFrame subclass, LocalFrameWithReturn, and use that
in all places except the constructor.

Amends d045e2d2b604152450a45846fc87cb53d89bf50a to address header review
comment.

Change-Id: Ie1391f0af47d1febaa8e01020f5b53b99287c89d
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit fbea1bdc180ccd8fe59d36b4eb0084e1bf264803)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>