QWindowsWindow: requestUpdate - Fix issue with recycled addresses
In CI we started seeing occasional asserts due to stale callbacks
calling deliverUpdateRequest(). The callbacks were not dropped as they
should have because the platform window was destroyed and recreated with
the same address. To mitigate this we now also verify that the callback
id matches. While it's technically possible that both the callbackId
and the recreated window has the same match, that's no longer a problem
because that also means a callback has already been requested, so
we'll just end-up dropping the last one that comes in.

Pick-to: 6.10
Change-Id: I8898c953789777e8ac457a21a6dbdbda935ac1d9
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 0f308aba520eb8ecafcb552258f53aca60740a9f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>