rcc: add another missing std::move()
Coverity complained that the 'pwd' variable could have been moved into
the third 'interpretResourceFile' parameter, but was instead copied.

It is correct. Add the std::move(), which doesn't impair readability,
because it appears near the end of the scope in which 'pwd' is
defined.

Amends the start of the public history.

Coverity-Id: 427453
Pick-to: 6.10 6.8
Change-Id: I292a33f2a561d61635ba34876abd8e9b21d839f7
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 15fd587515535ce93928e802c8ce3bcfa6f23247)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>