rcc: add a missing std::move()
Coverity complained that the 'compressed' variable could have been
moved into 'data', but was instead copied.

It is correct. Add the std::move(), which doesn't impair readability,
because it appears near the end of the scope in which 'compressed' is
defined.

Amends the start of the public history.

Coverity-Id: 464400
Pick-to: 6.10 6.8
Change-Id: I8bbcd17dff930dc92d7e4298891dc643c34383e3
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit b39a702e78d7708534261dde0e3d31cd0c2632ee)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>