QCommonStyle: sync SE_CheckBoxContents with SE_RadioButtonContents
The calculation in subElementRect for SE_CheckBoxContents and
SE_RadioButtonContents differed in 1px which resulted in a visual
misalignment of the icons/text when a Checkbox and a RadioButton were in
a layout below each other.
Therefore use the same calculation for both.

Pick-to: 6.10
Change-Id: I41bc3a4090f1cafc4a0bae5701fdbe8a512d6b1b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit c17ec87f0b54fb6ef0e490108d9fa37f1ddd2ae5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>