QCoreApplication/Linux: get the executable path using getauxval()
It's cheaper than qt_readlink() on /proc/self/exe because a) it avoids
the readlink() syscall and b) avoids allocating memory for the 8-bit
file name. This does mean we will not follow a rename of the executable,
but that was already a fragile behavior given that QCoreApplication
caches the resulting file name, and now does match what other OSes do.

The AT_EXECFN support was added to the Linux kernel in v2.6.27, so it's
plenty old for us (our oldest declared support in minimum-linux_p.h is
2.6.28). glibc added the getauxval() function in 2.16 (2012) and Bionic
for Android 4.3 (2013), both also plenty old for us, though we have a
configure-time check.

Drive-by fix indentation of the #endif.

Change-Id: If2f2df7b8fef6685ae24fffdfac7df16d4206819
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>